[Hack] Extract nested classes to avoid Truffle DSL bug
