+++ src/trufflesom/interpreter/nodes/specialized/AndBoolMessageNode.java
@@ -1,0 +1,17 @@
+package trufflesom.interpreter.nodes.specialized;
+
+import com.oracle.truffle.api.dsl.GenerateNodeFactory;
+import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.api.frame.VirtualFrame;
+
+import trufflesom.interpreter.nodes.nary.BinaryExpressionNode;
+
+
+@GenerateNodeFactory
+public abstract class AndBoolMessageNode extends BinaryExpressionNode {
+  @Specialization
+  public final boolean doAnd(final VirtualFrame frame, final boolean receiver,
+      final boolean argument) {
+    return receiver && argument;
+  }
+}
