+++ src/trufflesom/interpreter/nodes/specialized/OrBoolMessageNode.java
@@ -1,0 +1,17 @@
+package trufflesom.interpreter.nodes.specialized;
+
+import com.oracle.truffle.api.dsl.GenerateNodeFactory;
+import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.api.frame.VirtualFrame;
+
+import trufflesom.interpreter.nodes.nary.BinaryExpressionNode;
+
+
+@GenerateNodeFactory
+public abstract class OrBoolMessageNode extends BinaryExpressionNode {
+  @Specialization
+  public final boolean doOr(final VirtualFrame frame, final boolean receiver,
+      final boolean argument) {
+    return receiver || argument;
+  }
+}
