+++ src/trufflesom/interpreter/nodes/specialized/whileloops/WhilePrimitiveNode.java
@@ -3,7 +3,6 @@ package trufflesom.interpreter.nodes.specialized.whileloops;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.frame.VirtualFrame;
 
-import bd.primitives.Primitive;
 import trufflesom.interpreter.nodes.nary.BinaryExpressionNode.BinarySystemOperation;
 import trufflesom.vm.Universe;
 import trufflesom.vmobjects.SBlock;
@@ -32,23 +31,5 @@ public abstract class WhilePrimitiveNode extends BinarySystemOperation {
   protected SObject doWhileConditionally(final VirtualFrame frame,
       final SBlock loopCondition, final SBlock loopBody) {
     return (SObject) whileNode.executeEvaluated(frame, loopCondition, loopBody);
-  }
-
-  @Primitive(className = "Block", primitive = "whileTrue:", selector = "whileTrue:",
-      receiverType = SBlock.class, noWrapper = true)
-  // TODO: need to check for the second argument, check WhileSplzr
-  public abstract static class WhileTruePrimitiveNode extends WhilePrimitiveNode {
-    public WhileTruePrimitiveNode() {
-      super(true);
-    }
-  }
-
-  @Primitive(className = "Block", primitive = "whileFalse:", selector = "whileFalse:",
-      receiverType = SBlock.class, noWrapper = true)
-  // TODO: need to check for the second argument, check WhileSplzr
-  public abstract static class WhileFalsePrimitiveNode extends WhilePrimitiveNode {
-    public WhileFalsePrimitiveNode() {
-      super(false);
-    }
   }
 }
